travelai/pyproject.toml
[project]
name = "travelai"
version = "0.1.0"
description = "TravelAI using crewAI and Amadeus APIs for intelligent flight searches"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "reportlab>=4.3.1",
    "masumi-crewai>=0.1.26",
]

[project.scripts]
travelai = "travelai.main:run"
run_crew = "travelai.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
